+++ src/config.ini
@@ -57,7 +57,7 @@ frame_height = -1
 # The lower this setting is, the more dark frames are ignored
 dark_threshold = 50
 
-# The recorder to use. Can be either opencv, ffmpeg or pyv4l2.
+# The recorder to use. Can be either opencv (default), ffmpeg or pyv4l2.
 # Switching from the default opencv to ffmpeg can help with grayscale issues.
 recording_plugin = opencv
